Rename svg2pdf to svg2pdf.py
The 'extensionless' file is probably a linuxism that I'm not aware of. As a Windows user I had to go in and manually set this file's extension to `.py` in order to use it after installing from pip.
I am in no way pushing hard for this change, I just felt this was the most appropriate place to ask about it (didn't want to file an 'issue' or whatever). If this will be a "won't fix," would you be willing to spare a few minutes to educate me on the use-case of an extensionless script?
File extension grumbling aside - it works good, so thanks for building / maintaining this.